{{cookiecutter.project_slug}}/pyproject.toml
[tool.poetry]
name = "{{cookiecutter.project_slug}}"
version = "{{cookiecutter.project_version}}"
description = "{{cookiecutter.project_short_description}}"
authors = ["{{cookiecutter.full_name}} <{{cookiecutter.email}}>"]
readme = "README.md"
repository = "{{cookiecutter.project_git_path}}"

[tool.poetry.dependencies]
python = "3.8.10"
uvicorn = "0.12.1"
fastapi = "0.70.0"
ujson = "3.2.0"
email_validator = "1.1.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
psycopg2-binary = "2.8.6"
alembic = "1.4.3"
sqlalchemy = "1.3.19"
python-dotenv = "0.14.0"
typer = "0.3.2"
pytz = "^2020.1"
pydantic_sqlalchemy = "^0.0.7"
validate_email = "^1.3"
python-jose = "^3.2.0"
dictalchemy = "^0.1.2"
Jinja2 = "^2.11.3"
sentry-sdk = "0.19.0"
starlette-prometheus = "0.7.0"
piccolo-api = "^0.12.13"
netaddr = "^0.8.0"
requests = "^2.25.1"

[tool.poetry.dev-dependencies]
pytest = "6.1.0"
pytest-cov = "2.10.1"
flake8 = "3.8.3"
yapf = "^0.31.0"
rope = "^0.18.0"
bandit = "^1.7.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.yapf]
based_on_style = "google"
spaces_before_comment = 4
split_before_logical_operator = true